# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Movie Web App API
  description: |
    REST API for Movie Web App.

    ## Authentication
    Using **Google OAuth2** for login and **JWT Bearer Token** for accessing endpoints that require authentication.

    ## Roles
    - **guest** — Public access without authentication
    - **user** — User that has logged in via Google OAuth
    - **admin** — Administrator with full access
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://movie-ku-api.vercel.app
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from Google OAuth login
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
      required: [success, message]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
      required: [success, message]

    # --- Auth ---
...
    Genre:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      required: [id, name]

    # --- Movie Image ---
    MovieImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        image_type:
          type: string
          enum: [poster, backdrop]
        image_url:
          type: string
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true

    # --- Movie Video ---
    MovieVideo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        video_type:
          type: string
          enum: [trailer, teaser]
        site:
          type: string
          enum: [youtube, vimeo]
        video_key:
          type: string
        official:
          type: boolean

    # --- Movie ---
    Movie:
      type: object
      properties:
        id:
          type: string
          format: uuid
        api_id:
          type: string
          nullable: true
          description: TMDB API ID (null if user-created)
        source:
          type: string
          enum: [tmdb, user]
        title:
          type: string
        overview:
          type: string
          nullable: true
        release_date:
          type: string
          format: date
          nullable: true
        popularity:
          type: number
          format: float
          nullable: true
        rating:
          type: number
          format: float
          nullable: true
        is_featured:
          type: boolean
        status:
          type: string
          enum: [active, archived]
        created_by:
          type: string
          format: uuid
          nullable: true
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
        images:
          type: array
          items:
            $ref: "#/components/schemas/MovieImage"
        videos:
          type: array
          items:
            $ref: "#/components/schemas/MovieVideo"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, source, title, status, created_at, updated_at]
...
    AdminMovieCreateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        overview:
          type: string
        release_date:
          type: string
          format: date
        popularity:
          type: number
          format: float
        rating:
          type: number
          format: float
        is_featured:
          type: boolean
        status:
          type: string
          enum: [active, archived]
        genre_ids:
          type: array
          items:
            type: string
            format: uuid
      required: [title]
    AdminMovieUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        overview:
          type: string
        release_date:
          type: string
          format: date
        popularity:
          type: number
          format: float
        rating:
          type: number
          format: float
        is_featured:
          type: boolean
        status:
          type: string
          enum: [active, archived]
        genre_ids:
          type: array
          items:
            type: string
            format: uuid

    # --- Wishlist ---
...
    DashboardDateRange:
      type: object
      properties:
        start_date:
          type: string
          format: date
          example: "2026-02-05"
        end_date:
          type: string
          format: date
          example: "2026-03-07"
    DashboardSummary:
      type: object
      properties:
        total_movies:
          type: integer
          description: Movies created within date range
        total_users:
          type: integer
          description: Users registered within date range
        total_wishlists:
          type: integer
          description: Wishlists created within date range
        total_movies_all:
          type: integer
          description: Total movies (all time)
        total_users_all:
          type: integer
          description: Total users (all time)
        total_wishlists_all:
          type: integer
          description: Total wishlists (all time)
        top_genre:
          type: string
          nullable: true
          description: Most popular genre by movie count
        latest_movie:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            title:
              type: string
            created_at:
              type: string
              format: date-time
        average_rating:
          type: number
          format: float
          description: Average movie rating
    PieChartItem:
      type: object
      properties:
        label:
          type: string
        value:
          type: integer
        percentage:
          type: number
          format: float
          description: Percentage of this item relative to total (2 decimal places)
          example: 23.45
      required: [label, value, percentage]
    ColumnChartDailyItem:
      type: object
      properties:
        date:
          type: string
          format: date
        count:
          type: integer
      required: [date, count]
    ColumnChartRangeItem:
      type: object
      properties:
        range:
          type: string
          example: "6-8"
        count:
          type: integer
      required: [range, count]
    DashboardPieCharts:
      type: object
      properties:
        movies_by_genre:
          type: array
          items:
            $ref: "#/components/schemas/PieChartItem"
          description: Distribution of movies per genre (top 10)
        movies_by_source:
          type: array
          items:
            $ref: "#/components/schemas/PieChartItem"
          description: Movies by source (tmdb, user, admin)
        movies_by_status:
          type: array
          items:
            $ref: "#/components/schemas/PieChartItem"
          description: Movies by status (active, archived)
        sync_by_status:
          type: array
          items:
            $ref: "#/components/schemas/PieChartItem"
          description: Sync logs by status (success, failed)
    DashboardColumnCharts:
      type: object
      properties:
        movies_per_day:
          type: array
          items:
            $ref: "#/components/schemas/ColumnChartDailyItem"
          description: New movies added per day
        wishlists_per_day:
          type: array
          items:
            $ref: "#/components/schemas/ColumnChartDailyItem"
          description: Wishlists created per day
        users_per_day:
          type: array
          items:
            $ref: "#/components/schemas/ColumnChartDailyItem"
          description: New user registrations per day
        rating_distribution:
          type: array
          items:
            $ref: "#/components/schemas/ColumnChartRangeItem"
          description: Movies grouped by rating range (0-2, 2-4, 4-6, 6-8, 8-10)
    DashboardAnalytics:
      type: object
      properties:
        date_range:
          $ref: "#/components/schemas/DashboardDateRange"
        summary:
          $ref: "#/components/schemas/DashboardSummary"
        pie_charts:
          $ref: "#/components/schemas/DashboardPieCharts"
        column_charts:
          $ref: "#/components/schemas/DashboardColumnCharts"
      required: [date_range, summary, pie_charts, column_charts]

    # --- Admin User Management ---
    AdminUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]
        profile_picture:
          type: string
          nullable: true
        oauth_provider:
          type: string
          nullable: true
          description: 'google, github, or null if manually created'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, email, role, created_at, updated_at]
    AdminCreateUserRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
          maxLength: 255
        role:
          type: string
          enum: [user, admin]
          default: user
        profile_picture:
          type: string
          nullable: true
          maxLength: 500
      required: [name, email]
    AdminUpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        role:
          type: string
          enum: [user, admin]
          description: Admin cannot change their own role.
        profile_picture:
          type: string
          nullable: true
          maxLength: 500
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
        total_pages:
          type: integer
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Items per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        enum: [title, release_date, popularity, rating, created_at]
        default: created_at
      description: Field to sort by
    OrderParam:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort order
paths:
...
  /api/movies:
    get:
      tags: [Movies]
      summary: List all movies
      operationId: listMovies
      description: |
        Browse movies with support for search, filtering, sorting, and pagination.
        Accessible by anyone without authentication.
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by movie title (partial match)
        - name: genre_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by genre UUID
        - name: source
          in: query
          schema:
            type: string
            enum: [tmdb, user, admin]
          description: Filter by movie source
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived]
            default: active
          description: Filter by status
        - name: release_date_from
          in: query
          schema:
            type: string
            format: date
          description: Filter movies released on or after this date
        - name: release_date_to
          in: query
          schema:
            type: string
            format: date
          description: Filter movies released on or before this date
        - $ref: "#/components/parameters/SortParam"
        - $ref: "#/components/parameters/OrderParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: List of movies
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Movie"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
...
  /api/movies/{id}:
    get:
      tags: [Movies]
      summary: Get movie detail
      operationId: getMovieDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Movie detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Movie"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== USER MOVIES ====================
...
  /api/admin/dashboard:
    get:
      tags: [Admin Dashboard]
      summary: Get analytics dashboard
      operationId: getAdminDashboard
      description: |
        Returns full analytics dashboard data including summary cards,
        pie chart datasets (genre, source, status, sync distributions),
        and column chart datasets (daily movies/wishlists/users, rating distribution).
        All data is filtered by date range (defaults to last 30 days).
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
          description: Start of date range (YYYY-MM-DD). Defaults to 30 days ago.
          example: "2026-02-05"
        - name: end_date
          in: query
          schema:
            type: string
            format: date
          description: End of date range (YYYY-MM-DD). Defaults to today.
          example: "2026-03-07"
      responses:
        "200":
          description: Dashboard analytics
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DashboardAnalytics"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== ADMIN MOVIES ====================
  /api/admin/movies:
    get:
      tags: [Admin Movies]
      summary: List all movies (admin)
      operationId: listAdminMovies
      description: List all movies including archived. Admin can see all sources.
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
            enum: [tmdb, user, admin]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived]
        - $ref: "#/components/parameters/SortParam"
        - $ref: "#/components/parameters/OrderParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: All movies
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Movie"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Admin Movies]
      summary: Create movie (admin)
      operationId: createAdminMovie
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMovieCreateRequest"
      responses:
        "201":
          description: Movie created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Movie"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/movies/{id}:
    put:
      tags: [Admin Movies]
      summary: Update movie (admin)
      operationId: updateAdminMovie
      description: Admin can update any movie regardless of source or ownership.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMovieUpdateRequest"
      responses:
        "200":
          description: Movie updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Movie"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Admin Movies]
      summary: Delete movie (admin)
      operationId: deleteAdminMovie
      description: Soft-delete any movie. Admin can delete movies from any source.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Movie deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== ADMIN SYNC ====================
...
  /api/admin/users:
    get:
      tags: [Admin Users]
      summary: List all users
      operationId: listAdminUsers
      description: Returns paginated list of all users. Supports search by name/email and filter by role.
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Partial match on name or email
        - name: role
          in: query
          schema:
            type: string
            enum: [user, admin]
          description: Filter by role
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [name, email, created_at]
            default: created_at
        - name: order_by
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/AdminUserResponse"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Admin Users]
      summary: Create a new user
      operationId: createAdminUser
      description: |
        Admin manually creates a user account with any role (user or admin).
        No OAuth required — the account will have `oauth_provider: null`.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AdminUserResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/users/{id}:
    put:
      tags: [Admin Users]
      summary: Update a user
      operationId: updateAdminUser
      description: |
        Admin updates a user's name, role, or profile picture.

        **Restriction:** Admin cannot change their own role.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AdminUserResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only, or attempted self role change
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Admin Users]
      summary: Soft-delete a user
      operationId: deleteAdminUser
      description: |
        Soft-deletes a user account (sets `deleted_at`).

        **Restrictions:**
        - Only users with `role=user` can be deleted.
        - Admin accounts **cannot** be deleted by other admins.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — attempting to delete an admin account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"